requirements.txt
request
bs4
